Tidy up and add generic plan test
