+++ tests/unit_tests/optics/test_shutter_plans.py
@@ -1,4 +1,5 @@
 import pytest
+from blueapi.core import BlueskyContext
 from bluesky.run_engine import RunEngine
 from dodal.beamlines import i19_optics
 from dodal.devices.hutch_shutter import (
@@ -175,3 +176,28 @@ def test_operate_hutch_shutter_raises_error_if_hutch_invalid(
                 request_hutch, access_control_device, shutter_demand, expt_shutter
             )
         )
+
+
+def test_signature_of_shutter_plan(context: BlueskyContext):
+    plan = operate_shutter_plan
+    context.register_plan(plan)
+    assert plan.__name__ in context.plans
+
+    schema = context.plans["operate_shutter_plan"].model.model_json_schema()
+
+    # Plan
+    assert schema["title"] == "operate_shutter_plan"
+
+    # Check arguments
+    assert "experiment_hutch" in schema["properties"].keys()
+    assert "access_device" in schema["properties"].keys()
+    assert "shutter_demand" in schema["properties"].keys()
+    assert "shutter" in schema["properties"].keys()
+
+    # Check required arguments - shutter missing because injected
+    assert "shutter" not in schema["required"]
+    assert schema["required"] == [
+        "experiment_hutch",
+        "access_device",
+        "shutter_demand",
+    ]
